+++ ports/zephyr/common/memfault_platform_coredump_regions.c
@@ -15,41 +15,30 @@
 #include "memfault/ports/zephyr/version.h"
 
 #if MEMFAULT_ZEPHYR_VERSION_GT(2, 1)
-#include <arch/arm/aarch32/cortex_m/cmsis.h>
+  #include <arch/arm/aarch32/cortex_m/cmsis.h>
 #else
-#include <arch/arm/cortex_m/cmsis.h>
+  #include <arch/arm/cortex_m/cmsis.h>
 #endif
 
 #include "memfault/core/compiler.h"
 #include "memfault/core/math.h"
 #include "memfault/ports/zephyr/coredump.h"
 
-static sMfltCoredumpRegion s_coredump_regions[
-    MEMFAULT_COREDUMP_MAX_TASK_REGIONS
-    + 2 /* active stack(s) */
-    + 1 /* _kernel variable */
-    + 1 /* s_task_watermarks */
-    + 1 /* s_task_tcbs */
-#if CONFIG_MEMFAULT_COREDUMP_COLLECT_DATA_REGIONS
-    + 1
-#endif
-#if CONFIG_MEMFAULT_COREDUMP_COLLECT_BSS_REGIONS
-    + 1
-#endif
-    ];
-
-MEMFAULT_WEAK
-const sMfltCoredumpRegion *memfault_platform_coredump_get_regions(
-    const sCoredumpCrashInfo *crash_info, size_t *num_regions) {
+size_t memfault_zephyr_coredump_get_regions(const sCoredumpCrashInfo *crash_info,
+                                            sMfltCoredumpRegion *regions, size_t num_regions) {
+  // Check that regions is valid and has enough space to store all required regions
+  if (regions == NULL || num_regions < MEMFAULT_ZEPHYR_COREDUMP_REGIONS) {
+    return 0;
+  }
 
   const bool msp_was_active = (crash_info->exception_reg_state->exc_return & (1 << 2)) == 0;
-  int region_idx = 0;
+  size_t region_idx = 0;
 
   size_t stack_size_to_collect = memfault_platform_sanitize_address_range(
-        crash_info->stack_address, CONFIG_MEMFAULT_COREDUMP_STACK_SIZE_TO_COLLECT);
+    crash_info->stack_address, CONFIG_MEMFAULT_COREDUMP_STACK_SIZE_TO_COLLECT);
 
-  s_coredump_regions[region_idx] = MEMFAULT_COREDUMP_MEMORY_REGION_INIT(
-      crash_info->stack_address, stack_size_to_collect);
+  regions[region_idx] =
+    MEMFAULT_COREDUMP_MEMORY_REGION_INIT(crash_info->stack_address, stack_size_to_collect);
   region_idx++;
 
   if (msp_was_active) {
@@ -60,19 +49,15 @@ const sMfltCoredumpRegion *memfault_platform_coredump_get_regions(
     // exception frame that will have been stacked on it as well
     const uint32_t extra_stack_bytes = 128;
     stack_size_to_collect = memfault_platform_sanitize_address_range(
-        psp, CONFIG_MEMFAULT_COREDUMP_STACK_SIZE_TO_COLLECT + extra_stack_bytes);
-    s_coredump_regions[region_idx] = MEMFAULT_COREDUMP_MEMORY_REGION_INIT(
-        psp, stack_size_to_collect);
+      psp, CONFIG_MEMFAULT_COREDUMP_STACK_SIZE_TO_COLLECT + extra_stack_bytes);
+    regions[region_idx] = MEMFAULT_COREDUMP_MEMORY_REGION_INIT(psp, stack_size_to_collect);
     region_idx++;
   }
 
-  s_coredump_regions[region_idx] = MEMFAULT_COREDUMP_MEMORY_REGION_INIT(
-      &_kernel, sizeof(_kernel));
+  regions[region_idx] = MEMFAULT_COREDUMP_MEMORY_REGION_INIT(&_kernel, sizeof(_kernel));
   region_idx++;
 
-  region_idx += memfault_zephyr_get_task_regions(
-      &s_coredump_regions[region_idx],
-      MEMFAULT_ARRAY_SIZE(s_coredump_regions) - region_idx);
+  region_idx += memfault_zephyr_get_task_regions(&regions[region_idx], num_regions - region_idx);
 
   //
   // Now that we have captured all the task state, we will
@@ -81,15 +66,22 @@ const sMfltCoredumpRegion *memfault_platform_coredump_get_regions(
   //
 
 #if CONFIG_MEMFAULT_COREDUMP_COLLECT_DATA_REGIONS
-  region_idx +=  memfault_zephyr_get_data_regions(
-      &s_coredump_regions[region_idx], MEMFAULT_ARRAY_SIZE(s_coredump_regions) - region_idx);
+  region_idx += memfault_zephyr_get_data_regions(&regions[region_idx], num_regions - region_idx);
 #endif
 
 #if CONFIG_MEMFAULT_COREDUMP_COLLECT_BSS_REGIONS
-  region_idx +=  memfault_zephyr_get_bss_regions(
-      &s_coredump_regions[region_idx], MEMFAULT_ARRAY_SIZE(s_coredump_regions) - region_idx);
+  region_idx += memfault_zephyr_get_bss_regions(&regions[region_idx], num_regions - region_idx);
 #endif
 
-  *num_regions = region_idx;
+  return region_idx;
+}
+
+MEMFAULT_WEAK
+const sMfltCoredumpRegion *memfault_platform_coredump_get_regions(
+  const sCoredumpCrashInfo *crash_info, size_t *num_regions) {
+  static sMfltCoredumpRegion s_coredump_regions[MEMFAULT_ZEPHYR_COREDUMP_REGIONS];
+
+  *num_regions = memfault_zephyr_coredump_get_regions(crash_info, s_coredump_regions,
+                                                      MEMFAULT_ARRAY_SIZE(s_coredump_regions));
   return &s_coredump_regions[0];
 }