+++ tests/src/test_memfault_printf_attribute.cpp
@@ -1,11 +1,15 @@
+// Clang-format really hates this file, so disable it
+// clang-format off
 #include "CppUTest/MemoryLeakDetectorMallocMacros.h"
 #include "CppUTest/MemoryLeakDetectorNewMacros.h"
 #include "CppUTest/TestHarness.h"
 #include "CppUTestExt/MockSupport.h"
 
 extern "C" {
-  #define printf error_yolo
-  #include "memfault/core/compiler.h"
+#define printf error_yolo
+#if defined(printf)
+#include "memfault/core/compiler.h"
+#endif
 }
 
 MEMFAULT_PRINTF_LIKE_FUNC(1,2) void memfault_printf(const char *fmt, ...);
@@ -29,7 +33,7 @@ TEST_GROUP(MfltPrintfAttribute){
   }
 };
 
-TEST(MfltPrintfAttribute, Test_MfltHttpClientOverrides) {
+TEST(MfltPrintfAttribute, Test_PrintfImplementation) {
   // just call it. the real check is the compilation check though.
   memfault_printf("hello world\n");
 }